app/requirements.txt
fastapi>=0.115
uvicorn>=0.34
anthropic>=0.80
openai>=1.50
pyyaml>=6.0
itsdangerous>=2.1
python-jose[cryptography]>=3.3
httpx>=0.27
google-genai>=1.0
duckduckgo_search>=6.0
eval_type_backport>=0.2.0
